requirements.txt
Django==1.11.6
django-cors-headers==2.5.3
djangorestframework==3.9.3
pytz==2019.1
